Changed image
The image was changed as my Github README was updated.

The reason for the update is that few users have forked the repo
to use my README as a template, however, I found out that they
deleted their forks, and that is understandable as I find my
README very limited, specifically with the use of the table, the
whole area of the Markdown cannot be used fully.

As such I removed the table and completely changed the
layout of the README; I added some GIFs too to imply the
personality, hobbies, interests, etc. of the user.